requirements.txt
fastapi
uvicorn[standard]

# LangChain + Providers
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
langchain-openai
langgraph

# Embeddings
sentence-transformers

# Vector DB: Pinecone
langchain-pinecone
pinecone-client

# Ranking
rank_bm25
flashrank

# Env + Data
python-dotenv
datasets
pandas

# Evaluation
ragas

# Deployment 
redis
boto3
